+++ src/main/kotlin/org/deblock/flights/service/client/toughjet/ToughJetFlight.kt
@@ -1,12 +1,13 @@
 package org.deblock.flights.service.client.toughjet
 
+import java.math.BigDecimal
 import java.time.Instant
 
 data class ToughJetFlight(
     val carrier: String,
-    val basePrice: Double,
-    val tax: Double,
-    val discount: Double,
+    val basePrice: BigDecimal,
+    val tax: BigDecimal,
+    val discount: BigDecimal,
     val departureAirportName: String,
     val arrivalAirportName: String,
     val outboundDateTime: Instant,
